My brain is smooth.
It seems I nuked all of our previous work. It was replaced with a simply "Test.py" script that I was working on for minor modifications. given that this is the only script we have in the entire repo now, I guess it can't be named "test.py" anymore. 

We can pour one out for our fallen friends, but it's time to get back on the horse and commit some code! We've got a demo to put on in a week!